Fix: 修复 MySQL 绑定类型不匹配问题
将 score/total_count/win_count/id 变量类型从 unsigned long 改为 unsigned int
与 MYSQL_TYPE_LONG (32位) 匹配，解决分数变成随机值的问题

+++ source/include/user_table.hpp
@@ -431,14 +431,14 @@ inline bool UserTable::select_for_auth(const std::string& username, Json::Value&
     MYSQL_BIND result[7];
     memset(result, 0, sizeof(result));
 
-    unsigned long id;
+    unsigned int id;
     char username_buf[33];
     unsigned long username_len = 0;
     char password_hash_buf[513];
     unsigned long password_hash_len = 0;
-    unsigned long score;
-    unsigned long total_count;
-    unsigned long win_count;
+    unsigned int score;
+    unsigned int total_count;
+    unsigned int win_count;
     int status;
 
     bool is_null_false = false;
